refactor: minor repo reorg - minor change to move header into include/cplotlib/ - also simplifies cmake, no more need for 'virtual interface'
