+++ Config.cmake.in
@@ -5,5 +5,4 @@ include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
 check_required_components(@PROJECT_NAME@)
 
 include(CMakeFindDependencyMacro)
-find_dependency(nlohmann_json REQUIRED)
 find_dependency(Python3 REQUIRED)
